requirements.txt
python>=3.11
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
structlog>=24.1.0

